Pry apart forward and backward passes.
& cull some dependencies
