+++ crates/brush-android/Cargo.toml
@@ -12,7 +12,6 @@ egui.workspace = true
 eframe.workspace = true
 brush-app.path = "../brush-app"
 brush-ui.path = "../brush-ui"
-tokio_with_wasm = { workspace = true, features = ["rt"] }
 tokio = { workspace = true, features = ["io-util", "rt", "rt-multi-thread"] }
 
 rrfd.path = "../rrfd"
