+++ crates/brush-dataset/src/formats/nerfstudio.rs
@@ -13,8 +13,8 @@ use anyhow::Result;
 use async_fn_stream::try_fn_stream;
 use brush_render::camera::fov_to_focal;
 use brush_render::camera::{focal_to_fov, Camera};
-use brush_render::Backend;
 use brush_train::scene::SceneView;
+use burn::prelude::Backend;
 use std::future::Future;
 use std::path::Path;
 use std::sync::Arc;
