+++ crates/brush-prefix-sum/Cargo.toml
@@ -8,7 +8,6 @@ license.workspace = true
 [dependencies]
 brush-kernel.path = "../brush-kernel"
 burn.workspace = true
-burn-jit.workspace = true
 burn-wgpu.workspace = true
 naga_oil.workspace = true
 wgpu.workspace = true
