+++ crates/brush-render/src/tests/mod.rs
@@ -1,2 +1,1 @@
-mod reference;
 mod render;
