+++ crates/brush-render/src/tests/render.rs
@@ -1,15 +1,12 @@
-use crate::{camera::Camera, Backend};
+use crate::{camera::Camera, SplatForward};
 use assert_approx_eq::assert_approx_eq;
-use burn::{
-    backend::Autodiff,
-    tensor::{Tensor, TensorPrimitive},
-};
+use burn::tensor::{Tensor, TensorPrimitive};
 use burn_wgpu::{Wgpu, WgpuDevice};
 
-type DiffBack = Autodiff<Wgpu>;
+type Back = Wgpu;
 
-#[tokio::test]
-async fn renders_at_all() {
+#[test]
+fn renders_at_all() {
     // Check if rendering doesn't hard crash or anything.
     // These are some zero-sized gaussians, so we know
     // what the result should look like.
@@ -23,16 +20,16 @@ async fn renders_at_all() {
     let img_size = glam::uvec2(32, 32);
     let device = WgpuDevice::DefaultDevice;
     let num_points = 8;
-    let means = Tensor::<DiffBack, 2>::zeros([num_points, 3], &device);
-    let xy_dummy = Tensor::<DiffBack, 2>::zeros([num_points, 2], &device);
-    let log_scales = Tensor::<DiffBack, 2>::ones([num_points, 3], &device) * 2.0;
-    let quats: Tensor<DiffBack, 2> =
-        Tensor::<DiffBack, 1>::from_floats(glam::Quat::IDENTITY.to_array(), &device)
+    let means = Tensor::<Back, 2>::zeros([num_points, 3], &device);
+    let xy_dummy = Tensor::<Back, 2>::zeros([num_points, 2], &device);
+    let log_scales = Tensor::<Back, 2>::ones([num_points, 3], &device) * 2.0;
+    let quats: Tensor<Back, 2> =
+        Tensor::<Back, 1>::from_floats(glam::Quat::IDENTITY.to_array(), &device)
             .unsqueeze_dim(0)
             .repeat_dim(0, num_points);
-    let sh_coeffs = Tensor::<DiffBack, 3>::ones([num_points, 1, 3], &device);
-    let raw_opacity = Tensor::<DiffBack, 1>::zeros([num_points], &device);
-    let (output, aux) = DiffBack::render_splats(
+    let sh_coeffs = Tensor::<Back, 3>::ones([num_points, 1, 3], &device);
+    let raw_opacity = Tensor::<Back, 1>::zeros([num_points], &device);
+    let (output, aux) = <Back as SplatForward<Back>>::render_splats(
         &cam,
         img_size,
         means.into_primitive().tensor(),
@@ -45,7 +42,7 @@ async fn renders_at_all() {
     );
     aux.into_wrapped().debug_assert_valid();
 
-    let output: Tensor<DiffBack, 3> = Tensor::from_primitive(TensorPrimitive::Float(output));
+    let output: Tensor<Back, 3> = Tensor::from_primitive(TensorPrimitive::Float(output));
     let rgb = output.clone().slice([0..32, 0..32, 0..3]);
     let alpha = output.slice([0..32, 0..32, 3..4]);
     let rgb_mean = rgb.mean().to_data().as_slice::<f32>().expect("Wrong type")[0];
