+++ crates/brush-train/src/shaders/gather_grads.wgsl
