+++ crates/brush-train/test_cases/basic_case.safetensors
