+++ crates/brush-train/test_cases/crab.png
