+++ crates/brush-train/test_cases/tiny_case.safetensors
