+++ crates/sync-span/Cargo.toml
@@ -9,7 +9,6 @@ license.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 burn.workspace = true
-burn-wgpu.workspace = true
 
 [lints]
 workspace = true
