+++ deny.toml
@@ -16,6 +16,7 @@ targets = [
     { triple = "x86_64-pc-windows-msvc" },
     { triple = "x86_64-unknown-linux-gnu" },
     { triple = "x86_64-unknown-linux-musl" },
+    { triple = "aarch64-linux-android" },
 ]
 all-features = true
 
@@ -38,7 +39,6 @@ skip = [
     { name = 'itertools' },
     { name = 'derive-new' },
     { name = 'getrandom' },
-
 
     #
     { name = 'thiserror' },
